+++ src/Search.js
@@ -7,6 +7,7 @@ import makeQueryFunction from '@folio/stripes-components/util/makeQueryFunction'
 import SearchAndSort from '@folio/stripes-smart-components/lib/SearchAndSort';
 import { filterState } from '@folio/stripes-components/lib/FilterGroups';
 import ViewRecord from './ViewRecord';
+import redirectParams from './redirectParams';
 import packageInfo from '../package';
 import localIcon from '../icons/local-source.svg';
 import kbIcon from '../icons/generic.svg';
@@ -178,28 +179,14 @@ class Search extends React.Component {
   onSelectRow = (e, record) => {
     const logger = this.props.stripes.logger;
 
-    const obj = {};
-    if (record.source === 'kb') {
-      obj._path = `/eholdings/titles/${record.id}`;
-      obj.searchType = 'titles';
-      obj.q = _.get(this.props.resources, ['query', 'query']);
-      obj.query = null;
-    } else if (record.source === 'local') {
-      obj._path = `/inventory/view/${record.id}`;
-      obj.searchType = null;
-      obj.q = null;
-      // The 'query' parameter should already be correct
+    const obj = redirectParams(record, this.props.resources);
+    logger.log('action', `clicked ${record.id}, jumping to '${record.source}' version with obj`, obj);
+    if (obj) {
+      this.props.mutator.query.update(obj);
     } else {
       logger.log('action', `unsupported source '${record.source}': doing nothing`);
-      return true;
     }
 
-    logger.log('action', `clicked ${record.id}, jumping to '${record.source}' version with obj`, obj);
-
-    // TODO: reseting filters is a temp solution until
-    // https://issues.folio.org/browse/UISE-67 is fixed
-    obj.filters = '';
-    this.props.mutator.query.update(obj);
     return false;
   }